+++ Models/Vendedor.cs
@@ -37,7 +37,7 @@ namespace database_web.Models
         /// <summary>
         /// quantia de dinheiro do na conta do vendedor
         /// </summary>
-        public int dinheiro { get; set; }
+        public decimal dinheiro { get; set; }
 
         /// <summary>
         /// FK para um lista de anuncios, cujo vendedor fez
